agent-zero/requirements.txt
# Core dependencies
PyQt6>=6.4.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
anthropic>=0.8.1
mistralai>=0.0.7
groq>=0.4.2
together>=0.2.5
replicate>=0.22.0
stability-sdk>=0.8.5
elevenlabs>=0.2.27
google-cloud-texttospeech>=2.14.1
google-cloud-speech>=2.21.0
google-cloud-vision>=3.4.4
vertexai>=0.0.1

# Web and automation
selenium>=4.10.0
beautifulsoup4>=4.12.2
requests>=2.31.0
aiohttp>=3.9.1
playwright>=1.41.2
undetected-chromedriver>=3.5.4
webdriver-manager>=4.0.1
pyautogui>=0.9.54
keyboard>=0.13.5

# Security and credential management
keyring>=24.3.0
cryptography>=41.0.7
pyyaml>=6.0.1
bcrypt>=4.1.2
PyJWT>=2.8.0

# AI and ML
transformers>=4.37.2
torch>=2.1.2
sentencepiece>=0.1.99
accelerate>=0.27.2
huggingface-hub>=0.19.0
sentence-transformers>=2.2.2
bitsandbytes>=0.42.0
peft>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
safetensors>=0.4.1

# Image processing
Pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.3

# Utility packages
python-magic>=0.4.27
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0
webbrowser>=0.10.2
psutil>=5.9.8
py-cpuinfo>=9.0.0

# Additional functionality
duckduckgo-search>=4.4
html2text>=2020.1.16

# Error handling and retries
tenacity>=8.2.0
backoff>=1.11.1

# Voice Chat
pyaudio>=0.2.14
wave>=0.0.2 